Consolidate tabs in the gene page/annotation pages together
